refactor(ng-dev): move `BuiltPackageWithInfo` to `release/config` for use in prechecks
As part of our upcoming new customizable release prechecks, the
`BuiltPackageWithInfo` type should be exposed so that it can be used for
the custom check functions, and also to avoid a circular dependency
between the publish BUILD target and the upcoming pre-check BUILD target.
